#1822 oracle_undostat: rule for non space error count
The check oracle_undostat has a new rule for choosing the state when a
'cannot allocate another extent' in the Undo-Tablespace occurs. Oracle
counts this event as non space error count in v$undostat.

In previos versions this leads to a critical state. Now it is possible
to configure the state with a new rule.