#1824 FIX oracle_recovery_status: removed default values from Check
The old check has caused a lot of false messages. The reason is
the different behavior in Oracle Releases. There is no easy
solution for all Releases and a lot of users had disabled this
check.
Now the check only reports the current checkpoint time in the
database until a rule with levels is defined.

Please Note: This change only applies to new discovered services.