Discovery: never set node_info during inventory
When nodes of clusters are being service discovered do never set the
node information. This is because we cannot know wether the service
will actually be a clustered services before it is discovered. So
the inventory function will never know wether the service will be
part of a cluster or not.