#1918 ps: new option for checking the age of a process (on Linux)
The process check now shows the age of processes (the total elapsed time since
the start of a process) and also allows to set upper limits on this. Currently
this is only implemented in the Linux agent. You need to update that agent if
you want to make use of the new feature.