#1215 FIX Include check-functions from extra_sections in precompailed checks
Due a new feature it's no possible for a check to use agent sections from foreign checks. In some
cases this checks has a parse function. In rare cases the foreign check was not part check_table of the host.
This had been leading to a missing parse function.

Bug not affected the cmc microcore.