#1920 df: Linux filesystem check now supports displaying data reserved for root
Check_MK accounts space that is reserved for the root user on Linux and Unix
as used space. Usually 5% are being reserved for root when a new filesystem is
being created. With this new parameter option you can have Check_MK display
the current amount of reserved but yet unused space.