#1860 df and other filesystem checks: process total fs size as perfdata
Since the total size of a filesystem may change we store the total size
as additional performance data and plot it with rrdtools like the used space.
As a consequence another rrd database file is created for every filesystem,
and slightly more disk space will be used for the rrd data of your site.