#1938 FIX docsis_channels_upstream: fixed missing checks if channels had the same ChannelId
DOCSIS services used only the ChannelId as item name.
<ul>
<li>Upstream Channel 1</li>
<li>Upstream Channel 2</li>
<li>Upstream Channel 3</li>
</ul>

There are instances were this ChannelId is not unique, which causes the check
to loose some services. With this update the docsis_channels_upstream check now
uses an additional parameter - the interface index - to make this service name unique.

<ul>
<li>Upstream Channel 1337.1</li>
<li>Upstream Channel 1338.2</li>
<li>Upstream Channel 1343.3</li>
<li>Upstream Channel 1363.1</li>
<li>Upstream Channel 1373.2</li>
<li>Upstream Channel 1393.3</li>
</ul>

<b>Note:</b> This extended item name is only used if it is requires, means if there
are duplicate ChannelId entries for this device.