1991 emc_datadomain_temps: make configurable via WATO, add Perf-O-Meter
The check plugin <tt>emc_datadomain_temps</tt> has been converted to the
new standard implementation for all temperature checks. This means that
the check can now be configured via WATO with all available features
and also has a Perf-O-Meter. The price for this is that the name
of the service has been changed, e.g. from <tt>Temperature 1-1</tt> to
<tt>Temperature Chassis Ambient Enclosure 1</tt>. This makes it more
conveniant to write general rules, e.g. or all ambient sensors.