#1944 FIX hr_mem: no longer reports incorrect memory values when cached memory values are broken
Some devices reported negative used cached values. As a result the
cached memory was added to to the total memory - instead of substracted.
This has been fixed. If the value for the used cache seems to be broken,
the check uses the total cache value instead.
This new behaviour also applies to the memory buffers data.