#2099 Allow replication of Event Console rule and settings
The Event Console can now be configured in a distributed setup via WATO. This means
that if you have several Event Consoles then you can have the settings of the central
WATO site be replicated to the decentral Event Consoles and their changes being
activate by using the usual <i>Active Changes</i> mechanism.

In the site configuration there is a new checkbox that activates the
replication of the configuration.

Notes: In previous versions the configuration was also replicated, but not
activated.  And you could not select sites for which not to replicate the
configuration. The local Event Console is being handled like it was. This
means that changes are activated via the special button in the Event Console
settings page.