#2141 FIX Fix computation of explicit time ranges with time of day
When you entered an explicit time range (e.g. in the options for the
graph time range) and selected <i>Explicit time...</i> then 24 hours
would be added to the end time. This has been fixed. Note: the adding
of one day is still being done when selecting <i>Explicit date...</i>,
so that if you select the same start and end date the span is exactly
24 hours.