#1902 mk_oracle: Performance hint for RMAN checks
There are some environments with performance issues in RMAN check. There is
a known issue from Oracle regarding deleted objects in the controlfile of
the database. See Note 465378.1 in MyOracleSupport for details.

Another solution is adding a hint in the SQL statement. This werk adds the
function to insert the hint with a variable in <tr>mk_oracle.cfg</tr>. Please
add the following entry:
<tr>HINT_RMAN="+RULE"</tr>