#2176 One custom icon or action can be configured per process/service
It is now possible to configure a custom icon or action per process or
windows service. This is some sort of shortcut. In case of process
inventory, you had to configure two rules:

a) a process discovery rule
b) a rule to assign an icon to this service

With this new feature you can configure the icon to be used for the
found process right in the value of a process discovery rule.