#2202 FIX hr_fs: Remove Label:... from mount point, replace \ with /
This makes the filesystem check via SNMP kompatible with the one via
the Windows agent. Note: This change is incompatible since the item
names of the checks will change for Windows systems. You need to
rediscover in this case.

This does not affect Windows hosts that are monitored via the normal
Check_MK Agent.