#2255 FIX checkpoint_*: Fixed exception in scan function when sysDescr has less than 2 spaces
When running the service discovery on SNMP devices which have a sysDescr which conains
less than 2 spaces this lead into an exception. The error was present in 1.2.6 and
master branch but only showed up in the master branch, since the error handling in
the master branch has changed to be more verbose about such problems.